refactor with base model
